create utility folder
