+++ VandasPage/frontend/src/components/LevelChanger.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import EducationMaterial from './EducationMaterial';
 import DataContext from '../context/dataContext';
 import { useParams } from 'react-router';
-import AddEducationMaterial from './modallcontent/AddEducationMaterial';
+import AddEducationMaterial from './modallContent/AddEducationMaterial';
 import AuthContext from '../context/AuthProvider';
-import Dashboard from './Dashboard';
+import Dashboard from './utility/Dashboard';
 import { useContext } from 'react';
 import api from '../hooks/api';
 import { DragDropContext, Draggable } from 'react-beautiful-dnd';
